feat: simplified datetimerangequery
