+++ fts-core/src/models/datetime.rs
@@ -1,8 +1,26 @@
+/// A timestamped record of some value.
+///
+/// The interval for which the entity has this value is provided alongside
+/// the value itself.
+#[cfg_attr(
+    feature = "schemars",
+    derive(schemars::JsonSchema),
+    schemars(rename = "{Value}Record")
+)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+pub struct ValueRecord<DateTime, Value> {
+    /// The timestamp when this change occurred
+    pub valid_from: DateTime,
+    /// The timestamp when this change was superceded
+    pub valid_until: Option<DateTime>,
+    /// The component value
+    pub value: Value,
+}
+
 /// A query type for dealing with datetime ranges
 ///
 /// This structure enables API endpoints to accept parameters for filtering results
 /// based on time ranges with optional upper and lower bounds.
-#[derive(Debug, Clone, PartialEq)]
 #[cfg_attr(
     feature = "schemars",
     derive(schemars::JsonSchema),
@@ -23,7 +41,6 @@ pub struct DateTimeRangeQuery<DateTime> {
 ///
 /// This structure provides a standard format for returning time-based paginated results,
 /// including both the results and pagination metadata for retrieving the next page.
-#[derive(Debug, Clone, PartialEq)]
 #[cfg_attr(
     feature = "schemars",
     derive(schemars::JsonSchema),
@@ -32,7 +49,7 @@ pub struct DateTimeRangeQuery<DateTime> {
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct DateTimeRangeResponse<T, DateTime> {
     /// The collection of results matching the query
-    pub results: Vec<T>,
+    pub results: Vec<ValueRecord<DateTime, T>>,
 
     /// Optional pagination metadata for retrieving the next page of results.
     /// If present, indicates there are more results available.
